EventSequence now has the Wait and Gap functions.
While working on the first slice of the video, I noticed that it was very cumbersome to specify the time until an event starts on the event itself. Instead, there are now Wait and Gap functions. Wait will add a blank event that does not start until the previous event ends. Gap will just add a buffer of time. This gives us an easy way to wait for events to finish before others begin and create space between them.

Also, The Add event function just takes in a single struct that specifies all of the event properties. This just seemed like a nicer way to pass in the data because it's a bit more explicit about what each parameter is. It was something I was having trouble keeping track of even though I knew exactly what parameters were what.