Objects are now being used instead of MemberIds.
The reason I didn't start by doing this is because Objects hold onto a space pointer that tracks where the object is contained. By using Objects instead of MembmerIds, we are creating a crap ton of copies of the same space pointer because each object holds onto a copy of that pointer.

I could try to resolve this somehow, but really, Objects are there to give users something easier to interact. That ease of use just comes with a bit of a sacrifice, and tbh, I'm ok with that because it means I can type less.