The sequence can now be forward and reverse scrubbed.
This is extremely useful for being able to jump to any part of the animation without having to wait for the sequence to reach a particular section. The sequence now uses absolute times instead of relative times, which made implementing this much easier. Crucially, the user does not need to worry about absolute times. The Sequence extracts absolute times given the relative timings supplied by the user.

Some of the event functions were updated to account for scrubbing. When t is at 0, the function must put the animation at its starting state. When t is at 1, the function must result in the animation's end state.