Sequence Begin/End Crossings and Rod Colors
There exists certain state that needs to be established when an event is entered or exited from its beginning/end. Establishing this state with only the lerp function was challenging and not very explicit, so I decided to add the Begin and End functions to make the process of changing state on event borders easier.

Doing that worked. Changing the colors of the rods at event borders was much easier once I was able to use the begin and end functions. Before that, I was using multiple events to perform the state changes and it was incredibly hard to reason about.