Event Begin Bugfix
When mTimePassed was 0, the Begin function of the first event was not called. Strangely, this behaviour only occurred after performing an initial scrub up and than scrubbing back down to the start of the seqeunce. Regardless, this seems like the right change because it conforms to the idea that all event borders are treated inclusively.