Camera Framing
The point clouds are now appropriately centered and the camera rotates around the point cloud as the quick hull animation plays out. This required being able to add `Sequence::DiscreteEvent` to a sequence. Without it, it was not possible to specify start and end times for an event. Because the camera animation happens over the course of all other events and not necessarily after a certain event, being able to define start and end times explicitly made establishing the correct camera animations easier.

A note on those camera animations: specifically the fast rotation that occurs when a point cloud grows into the scene and when the convex hull fades from the scene. A goal is that clicking at some time t on the scrub bar will always produce the same output frame. Because the desired camera property to control is velocity, ensuring the camera positions would always be the same proved challenging. The CameraInfo::StartTheta and CameraInfo::EndTheta functions were the solution to this problem. They express the integral with respect to the camera's angular velocity. mWs is the starting angular velocity and mWc is the angular velocity that is lerped to with a QuadIn (StartTheta) or QuadOut (EndTheta) ease type.